+++ src/partials/games/FeaturedCategories.svelte
