+++ src/partials/games/SearchBar.svelte
