+++ src/partials/gift-cards/Sidebar.svelte
